Moved StorageEmulator to Utils
