+++ src/OrleansTestingHost/OrleansTestingHost.csproj
@@ -55,7 +55,7 @@
     <Compile Include="OrleansTestSecrets.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Utils\AsyncResultHandle.cs" />
-    <Compile Include="StorageEmulator.cs" />
+    <Compile Include="Utils\StorageEmulator.cs" />
     <Compile Include="StorageTestConstants.cs" />
     <Compile Include="SiloHandle.cs" />
     <Compile Include="TestingClientOptions.cs" />
